+++ controllers/dogu_controller_int_test.go
@@ -1,13 +1,14 @@
 //go:build k8s_integration
 // +build k8s_integration
 
-package controllers_test
+package controllers
 
 import (
 	"context"
 	_ "embed"
 	"fmt"
 	cesmocks "github.com/cloudogu/cesapp-lib/registry/mocks"
+	cesremotemocks "github.com/cloudogu/cesapp-lib/remote/mocks"
 	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
 	"github.com/cloudogu/k8s-dogu-operator/controllers/mocks"
 	. "github.com/onsi/ginkgo"
@@ -18,9 +19,16 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"testing"
 )
 
 var _ = Describe("Dogu Controller", func() {
+	t := &testing.T{}
+	ldapCr := readTestDataLdapCr(t)
+	redmineCr := readTestDataRedmineCr(t)
+	imageConfig := readTestDataImageConfig(t)
+	ldapDogu := readTestDataLdapDogu(t)
+	redmineDogu := readTestDataRedmineDogu(t)
 
 	ldapCr.Namespace = "default"
 	ldapCr.ResourceVersion = ""
@@ -33,13 +41,9 @@ var _ = Describe("Dogu Controller", func() {
 	ctx := context.TODO()
 	ImageRegistryMock = mocks.ImageRegistry{}
 	ImageRegistryMock.Mock.On("PullImageConfig", mock.Anything, mock.Anything).Return(imageConfig, nil)
-	DoguRegistryMock = mocks.DoguRegistry{}
-	DoguRegistryMock.Mock.On("GetDogu", mock.MatchedBy(func(dogu *k8sv1.Dogu) bool {
-		return dogu.Name == ldapCr.Name
-	})).Return(ldapDogu, nil)
-	DoguRegistryMock.Mock.On("GetDogu", mock.MatchedBy(func(dogu *k8sv1.Dogu) bool {
-		return dogu.Name == redmineCr.Name
-	})).Return(redmineDogu, nil)
+	DoguRemoteRegistryMock = cesremotemocks.Registry{}
+	DoguRemoteRegistryMock.Mock.On("Get", "official/ldap").Return(ldapDogu, nil)
+	DoguRemoteRegistryMock.Mock.On("Get", "official/redmine").Return(redmineDogu, nil)
 
 	EtcdDoguRegistry = cesmocks.DoguRegistry{}
 	EtcdDoguRegistry.Mock.On("Get", "postgresql").Return(nil, fmt.Errorf("not installed"))