+++ controllers/suite_int_test.go
@@ -1,24 +1,28 @@
 //go:build k8s_integration
 // +build k8s_integration
 
-package controllers_test
+package controllers
 
 import (
 	"context"
 	_ "embed"
+	"github.com/bombsimon/logrusr/v2"
+	"github.com/cloudogu/k8s-dogu-operator/controllers/config"
+	"github.com/cloudogu/k8s-dogu-operator/controllers/limit"
+	resourceMocks "github.com/cloudogu/k8s-dogu-operator/controllers/resource/mocks"
+	"github.com/onsi/ginkgo"
+	"github.com/onsi/gomega"
+	"os"
 	"path/filepath"
 	"testing"
 	"time"
 
-	"github.com/bombsimon/logrusr/v2"
 	"github.com/cloudogu/cesapp-lib/core"
 	cesmocks "github.com/cloudogu/cesapp-lib/registry/mocks"
-	"github.com/cloudogu/k8s-dogu-operator/controllers"
+	cesremotemocks "github.com/cloudogu/cesapp-lib/remote/mocks"
 	"github.com/cloudogu/k8s-dogu-operator/controllers/dependency"
 	"github.com/cloudogu/k8s-dogu-operator/controllers/mocks"
 	"github.com/cloudogu/k8s-dogu-operator/controllers/resource"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/mock"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -42,7 +46,7 @@ var cancel context.CancelFunc
 var ImageRegistryMock mocks.ImageRegistry
 
 // Used in other integration tests
-var DoguRegistryMock mocks.DoguRegistry
+var DoguRemoteRegistryMock cesremotemocks.Registry
 
 // Used in other integration tests
 var EtcdDoguRegistry cesmocks.DoguRegistry
@@ -51,39 +55,46 @@ const TimeoutInterval = time.Second * 10
 const PollingInterval = time.Second * 1
 
 func TestAPIs(t *testing.T) {
-	RegisterFailHandler(Fail)
+	gomega.RegisterFailHandler(ginkgo.Fail)
 
-	RunSpecsWithDefaultAndCustomReporters(t,
+	ginkgo.RunSpecsWithDefaultAndCustomReporters(t,
 		"Controller Suite",
-		[]Reporter{printer.NewlineReporter{}})
+		[]ginkgo.Reporter{printer.NewlineReporter{}})
 }
 
-var _ = BeforeSuite(func() {
+var _ = ginkgo.BeforeSuite(func() {
+	// We need to ensure that the development stage flag is not passed by our makefiles to prevent the dogu operator
+	// from running in the developing mode. The developing mode changes some operator behaviour. Our integration test
+	// aim to test the production functionality of the operator.
+	err := os.Unsetenv(config.StageEnvironmentVariable)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	err = os.Setenv(config.StageEnvironmentVariable, config.StageProduction)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	config.Stage = config.StageProduction
+
 	logf.SetLogger(logrusr.New(logrus.New()))
 
 	var ctx context.Context
 	ctx, cancel = context.WithCancel(context.TODO())
 
-	By("bootstrapping test environment")
+	ginkgo.By("bootstrapping test environment")
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths:     []string{filepath.Join("..", "config", "crd", "bases")},
 		ErrorIfCRDPathMissing: true,
 	}
 
 	cfg, err := testEnv.Start()
-	Expect(err).NotTo(HaveOccurred())
-	Expect(cfg).NotTo(BeNil())
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	gomega.Expect(cfg).NotTo(gomega.BeNil())
 
 	err = k8sv1.AddToScheme(scheme.Scheme)
-	Expect(err).NotTo(HaveOccurred())
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
 	//+kubebuilder:scaffold:scheme
 	k8sManager, err := ctrl.NewManager(cfg, ctrl.Options{
 		Scheme: scheme.Scheme,
 	})
-	Expect(err).ToNot(HaveOccurred())
-
-	resourceGenerator := resource.NewResourceGenerator(k8sManager.GetScheme())
+	gomega.Expect(err).ToNot(gomega.HaveOccurred())
 
 	doguConfigurationContext := &cesmocks.ConfigurationContext{}
 	doguConfigurationContext.On("Set", mock.Anything, mock.Anything).Return(nil)
@@ -92,13 +103,18 @@ var _ = BeforeSuite(func() {
 	globalConfigurationContext := &cesmocks.ConfigurationContext{}
 	globalConfigurationContext.On("Get", "key_provider").Return("", nil)
 
-	CesRegistryMock := cesmocks.Registry{}
+	CesRegistryMock := &cesmocks.Registry{}
 	CesRegistryMock.On("DoguRegistry").Return(&EtcdDoguRegistry)
 	CesRegistryMock.On("DoguConfig", mock.Anything).Return(doguConfigurationContext)
 	CesRegistryMock.On("GlobalConfig").Return(globalConfigurationContext)
 
+	limitPatcher := &resourceMocks.LimitPatcher{}
+	limitPatcher.On("RetrievePodLimits", mock.Anything).Return(limit.DoguLimits{}, nil)
+	limitPatcher.On("PatchDeployment", mock.Anything, mock.Anything).Return(nil)
+	resourceGenerator := resource.NewResourceGenerator(k8sManager.GetScheme(), limitPatcher)
+
 	version, err := core.ParseVersion("0.0.0")
-	Expect(err).ToNot(HaveOccurred())
+	gomega.Expect(err).ToNot(gomega.HaveOccurred())
 
 	dependencyValidator := dependency.NewCompositeDependencyValidator(&version, &EtcdDoguRegistry)
 	serviceAccountCreator := &mocks.ServiceAccountCreator{}
@@ -109,45 +125,60 @@ var _ = BeforeSuite(func() {
 	doguSecretHandler := &mocks.DoguSecretsHandler{}
 	doguSecretHandler.On("WriteDoguSecretsToRegistry", mock.Anything, mock.Anything).Return(nil)
 
-	doguRegistrator := controllers.NewCESDoguRegistrator(k8sManager.GetClient(), &CesRegistryMock, resourceGenerator)
+	doguRegistrator := newCESDoguRegistrator(k8sManager.GetClient(), CesRegistryMock, resourceGenerator)
 
 	yamlResult := make(map[string]string, 0)
-	fileExtract := &mockFileExtractor{}
+	fileExtract := &mocks.FileExtractor{}
 	fileExtract.On("ExtractK8sResourcesFromContainer", mock.Anything, mock.Anything, mock.Anything).Return(yamlResult, nil)
-	applyClient := &mockApplier{}
+	applyClient := &mocks.Applier{}
 	applyClient.On("Apply", mock.Anything, mock.Anything).Return(nil)
 
-	doguManager := &controllers.DoguManager{
-		Client:                k8sManager.GetClient(),
-		Scheme:                k8sManager.GetScheme(),
-		ResourceGenerator:     resourceGenerator,
-		DoguRemoteRegistry:    &DoguRegistryMock,
-		DoguLocalRegistry:     &EtcdDoguRegistry,
-		ImageRegistry:         &ImageRegistryMock,
-		DoguRegistrator:       doguRegistrator,
-		DependencyValidator:   dependencyValidator,
-		ServiceAccountCreator: serviceAccountCreator,
-		ServiceAccountRemover: serviceAccountRemover,
-		DoguSecretHandler:     doguSecretHandler,
-		Applier:               applyClient,
-		FileExtractor:         fileExtract,
+	installManager := &doguInstallManager{
+		client:                k8sManager.GetClient(),
+		scheme:                k8sManager.GetScheme(),
+		resourceGenerator:     resourceGenerator,
+		doguRemoteRegistry:    &DoguRemoteRegistryMock,
+		doguLocalRegistry:     &EtcdDoguRegistry,
+		imageRegistry:         &ImageRegistryMock,
+		doguRegistrator:       doguRegistrator,
+		dependencyValidator:   dependencyValidator,
+		serviceAccountCreator: serviceAccountCreator,
+		doguSecretHandler:     doguSecretHandler,
+		applier:               applyClient,
+		fileExtractor:         fileExtract,
 	}
 
-	err = controllers.NewDoguReconciler(k8sManager.GetClient(), k8sManager.GetScheme(), doguManager).SetupWithManager(k8sManager)
-	Expect(err).ToNot(HaveOccurred())
+	deleteManager := &doguDeleteManager{
+		client:                k8sManager.GetClient(),
+		scheme:                k8sManager.GetScheme(),
+		doguLocalRegistry:     &EtcdDoguRegistry,
+		imageRegistry:         &ImageRegistryMock,
+		doguRegistrator:       doguRegistrator,
+		serviceAccountRemover: serviceAccountRemover,
+		doguSecretHandler:     doguSecretHandler,
+	}
+
+	doguManager := &DoguManager{
+		scheme:         k8sManager.GetScheme(),
+		installManager: installManager,
+		deleteManager:  deleteManager,
+	}
+
+	err = NewDoguReconciler(k8sManager.GetClient(), k8sManager.GetScheme(), doguManager).SetupWithManager(k8sManager)
+	gomega.Expect(err).ToNot(gomega.HaveOccurred())
 
 	go func() {
 		err = k8sManager.Start(ctx)
-		Expect(err).ToNot(HaveOccurred())
+		gomega.Expect(err).ToNot(gomega.HaveOccurred())
 	}()
 
 	k8sClient = k8sManager.GetClient()
-	Expect(k8sClient).ToNot(BeNil())
+	gomega.Expect(k8sClient).ToNot(gomega.BeNil())
 }, 60)
 
-var _ = AfterSuite(func() {
+var _ = ginkgo.AfterSuite(func() {
 	cancel()
-	By("tearing down the test environment")
+	ginkgo.By("tearing down the test environment")
 	err := testEnv.Stop()
-	Expect(err).NotTo(HaveOccurred())
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 })